Copy changes, added conditions for rendering, refactored into separate components and files
