+++ src/components/account/SecondPillarUpsell/SecondPillarUpsell.scss
@@ -1,4 +1,4 @@
-@import "../variables";
+@import "../../variables";
 
 .badge-pink {
   background: rgba(255, 72, 0, 0.20);
